improv(booking): use stripe env var as per docs
