+++ src/frontend/store/bookings/actions.js
@@ -2,16 +2,13 @@ import Booking from "../../shared/models/BookingClass";
 import Flight from "../../shared/models/FlightClass"; // eslint-disable-line
 // @ts-ignore
 import { Loading } from "quasar";
-import axios from "axios";
+import { processPayment } from "./payment";
 
 import { API, graphqlOperation } from "aws-amplify";
 import {
   processBooking as processBookingMutation,
   getBookingByStatus
 } from "./graphql";
-
-const paymentEndpoint =
-  process.env.VUE_APP_PaymentChargeUrl || "no payment gateway endpoint set";
 
 /**
  *
@@ -128,7 +125,7 @@ export async function createBooking(
     const customerEmail = rootState.profile.user.attributes.email;
 
     console.info(
-      `Processing payment before proceeding to booking for flight ${outboundFlight}`
+      `Processing payment before proceeding to book flight ${outboundFlight}`
     );
     let chargeToken = await processPayment({
       paymentToken,
@@ -136,112 +133,21 @@ export async function createBooking(
       customerEmail
     });
 
-    Loading.show({
-      message: "Payment authorized successfully..."
-    });
-
     console.info(
       `Creating booking with token ${chargeToken} for flight ${outboundFlight}`
     );
-    let bookingProcessId = await processBooking({
-      chargeToken,
-      outboundFlight
-    });
 
-    console.log(`Booking Id: ${bookingProcessId}`);
-    console.groupEnd();
-    return bookingProcessId;
-  } catch (err) {
-    throw err;
-  }
-}
-
-/**
- *
- * Process Payment function - processPayment calls Payment endpoint to pre-authorize charge upon tokenized payment details
- *
- * @param {object} obj - Object containing params to process payment
- * @param {object} obj.paymentToken - Tokenized payment info
- * @param {object} obj.paymentToken.details - Tokenized payment details including last4, id, etc.
- * @param {object} obj.paymentToken.id - Payment token
- * @param {Flight} obj.outboundFlight - Outbound flight
- * @param {string} obj.customerEmail - Customer Email address for payment notification
- * @returns {promise} - Promise representing whether payment was successfully pre-authorized
- * @example
- *   let chargeToken = await processPayment({
- *      paymentToken,
- *      outboundFlight,
- *      customerEmail
- *   });
- */
-async function processPayment({ paymentToken, outboundFlight, customerEmail }) {
-  console.group("store/bookings/actions/processPayment");
-  Loading.show({
-    message: "Charging a pre-authorization..."
-  });
-
-  if (!paymentToken) throw "Invalid payment token";
-
-  const chargeData = {
-    amount: outboundFlight.ticketPrice,
-    currency: outboundFlight.ticketCurrency,
-    stripeToken: paymentToken.details.id,
-    description: `Payment by ${customerEmail}`,
-    email: customerEmail
-  };
-
-  console.log("Charge data to be processed");
-  console.log(chargeData);
-  try {
-    const data = await axios.post(paymentEndpoint, chargeData);
-    const {
-      data: {
-        createdCharge: { id: chargeId }
-      }
-    } = data;
-
-    console.groupEnd();
-    return chargeId;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-}
-
-/**
- *
- * Process Booking function - processBooking uses processBooking mutation to kick off an async Booking Workflow that ultimatelly reserves flight seat, creates a booking reference, collect payment, etc.
- *
- * @param {object} obj - Object containing params to process payment
- * @param {string} obj.chargeToken - Pre-authorized payment token
- * @param {Flight} obj.outboundFlight - Outbound flight
- * @returns {promise} - Promise representing whether Booking Workflow was successfully initiated
- * @example
- *   const {
- *   // @ts-ignore
- *     data: {
- *      processBooking: { id: bookingProcessId }
- *     }
- *   } = await API.graphql(
- *     graphqlOperation(processBookingMutation, processBookingInput)
- *   );
- */
-async function processBooking({ chargeToken, outboundFlight }) {
-  console.group("store/bookings/actions/processBooking");
-  const processBookingInput = {
-    input: {
-      paymentToken: chargeToken,
-      bookingOutboundFlightId: outboundFlight.id
-    }
-  };
-
-  try {
     Loading.show({
       message: "Creating a new booking..."
     });
 
-    console.log("Booking data to be processed");
-    console.log(processBookingInput);
+    const processBookingInput = {
+      input: {
+        paymentToken: chargeToken,
+        bookingOutboundFlightId: outboundFlight.id
+      }
+    };
+
     const {
       // @ts-ignore
       data: {
@@ -251,9 +157,10 @@ async function processBooking({ chargeToken, outboundFlight }) {
       graphqlOperation(processBookingMutation, processBookingInput)
     );
 
+    console.log(`Booking Id: ${bookingProcessId}`);
+    console.groupEnd();
     return bookingProcessId;
   } catch (err) {
-    console.error(err);
     throw err;
   }
 }
